refactored to DRF-3
