+++ example/api/resources/identity.py
@@ -1,8 +1,8 @@
 from django.contrib.auth import models as auth_models
-from rest_framework import viewsets, generics, renderers, parsers
-from rest_framework.decorators import list_route, detail_route
-from rest_framework.response import Response
-from rest_framework_ember import mixins, utils
+from rest_framework_3 import viewsets, generics, renderers, parsers
+from rest_framework_3.decorators import list_route, detail_route
+from rest_framework_3.response import Response
+from rest_framework_ember_3 import mixins, utils
 from ..serializers.identity import IdentitySerializer
 from ..serializers.post import PostSerializer
