+++ example/api/serializers/post.py
@@ -1,4 +1,4 @@
-from rest_framework import serializers
+from rest_framework_3 import serializers
 
 
 class PostSerializer(serializers.Serializer):
