+++ setup.py
@@ -11,12 +11,12 @@ def get_readme():
 
 
 setup(
-    name='rest_framework_ember',
+    name='rest_framework_ember_3',
     version='1.3.2',
     description="A Django Rest Framework adapter that provides Ember Data \
         support. When jsonapi.org reaches 1.0 this adapter plans to adopt it.",
     long_description=get_readme(),
-    url='https://github.com/django-json-api/rest_framework_ember',
+    url='https://github.com/django-json-api/rest_framework_ember_3',
     license='BSD',
     keywords="EmberJS Ember Data Django REST",
     packages=find_packages(),
